// web/lib/api/generated/model/bucket.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface Bucket {
  name: string;
  /** @pattern ^-?\d{0,12}(?:\.\d{0,2})?$ */
  value: string;
  products: number;
  missing_cost: number;
}